pyproject.toml
[project]
name = "debezium-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kafka-python>=2.0.2",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
]
